backend/requirements.txt
# Core dependencies
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.9.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyYAML>=6.0

# Database and storage
aiofiles>=0.8.0
redis>=4.2.0
qdrant-client>=1.1.1
motor>=3.1.1

# LLM and ML dependencies
openai>=1.3.0
anthropic>=0.3.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.0.2

# Observability and monitoring
opentelemetry-api>=1.11.1
opentelemetry-sdk>=1.11.1
opentelemetry-instrumentation-fastapi>=0.30b1
opentelemetry-exporter-jaeger>=1.11.1
prometheus-client>=0.14.1
psutil>=5.9.0  # For system monitoring
py-spy>=0.3.12  # For profiling

# Distributed computing
grpcio>=1.44.0
protobuf>=3.19.4
aiohttp>=3.8.1

# Testing
pytest>=7.1.1
pytest-asyncio>=0.18.3
pytest-cov>=3.0.0
httpx>=0.22.0

# Development
black>=22.3.0
isort>=5.10.1
mypy>=0.950
pylint>=2.13.8